requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.6
Django==3.2.6
django-crispy-forms==1.12.0
django-multiselectfield==0.1.12
django-phone-field==1.8.1
idna==3.2
Pillow==8.3.2
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twilio==7.0.0
urllib3==1.26.7